CRUD WELL DONE(Mongo + Redis)

+++ mongodbDemo/src/main/java/ru/elmanov/mongodb/demo/controller/UserController.java
@@ -1,12 +1,14 @@
 package ru.elmanov.mongodb.demo.controller;
 
 import jakarta.annotation.PostConstruct;
+import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.*;
 import ru.elmanov.mongodb.demo.model.Car;
 import ru.elmanov.mongodb.demo.model.User;
+import ru.elmanov.mongodb.demo.model.enums.Sex;
 import ru.elmanov.mongodb.demo.repository.redis.UserRedisRepository;
 import ru.elmanov.mongodb.demo.service.CarService;
 import ru.elmanov.mongodb.demo.service.UserService;
@@ -25,15 +27,15 @@ public class UserController {
 
     @PostConstruct
     public void init() {
-        var car = new Car("Volvo");
+//        var car = new Car("Volvo");
 //        var insertedCar = carService.createCar(car);
-        User user = new User();
-        user.setAge(12);
-        user.setName("Tom");
-        user.setEmail("asd@sdf.rt");
-        user.setSex(User.Sex.MALE);
+//        User user = new User();
+//        user.setAge(12);
+//        user.setName("Tom");
+//        user.setEmail("asd@sdf.rt");
+//        user.setSex(Sex.MALE);
 //        user.setCar(insertedCar);
-        userService.createUser(user);
+//        userService.createUser(user);
     }
 
     @GetMapping
@@ -57,21 +59,28 @@ public class UserController {
     }
 
     @PostMapping
-    public ResponseEntity<?> createUser(@RequestBody User user) {
+    public ResponseEntity<?> createUser(@RequestBody @NonNull User user) {
         var createdUser = userService.createUser(user);
-        userRedisRepository.add(createdUser);
-        return ResponseEntity.ok().body(createdUser);
+
+        if (Objects.nonNull(createdUser)) {
+            userRedisRepository.add(createdUser);
+            return ResponseEntity.ok().body(createdUser);
+        }
+
+        return ResponseEntity.ok().body("Already exists");
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Integer> deleteUser(@PathVariable String id) {
+    public ResponseEntity<Integer> deleteUser(@PathVariable @NonNull String id) {
         var countDeleted = userService.deleteUser(id);
-        userRedisRepository.delete(id);
+        if (Objects.nonNull(countDeleted)) {
+            userRedisRepository.delete(id);
+        }
         return ResponseEntity.ok().body(countDeleted);
     }
 
     @PutMapping
-    public ResponseEntity<?> updateUser(@RequestBody User user) {
+    public ResponseEntity<?> updateUser(@RequestBody @NonNull User user) {
         return ResponseEntity.ok().body(userService.updateUser(user));
     }
 }
